Modified domain layer models, data layer DTOs to remove unused properties and unnecessary nullability. Fixed errors in code due to these changes.

+++ app/src/main/java/com/andrii_a/walleria/ui/common/NavigationArgumentWrappers.kt
@@ -12,7 +12,7 @@ value class UserNickname(val value: String)
 data class CollectionInfo(
     val idAsString: String,
     val title: String,
-    val totalPhotos: Int,
+    val totalPhotos: Long,
     val userNickname: String,
     val userFullName: String,
     val description: String?,
