+++ app/src/main/java/com/andrii_a/walleria/ui/photo_details/components/PhotoDetailsComponents.kt
@@ -491,7 +491,7 @@ fun StatsRow(
 fun RelatedCollectionsItem(
     title: String,
     coverPhotoUrl: String,
-    totalPhotos: Int,
+    totalPhotos: Long,
     modifier: Modifier = Modifier
 ) {
     Card(
